+++ back/routes/users.routes.js
@@ -1,6 +1,7 @@
 const express = require('express')
 const router = express.Router()
-const controllersthomas = require('../controllersthomas/controllers')
+const controllersthomas = require('../controllers/users.controllers')
 
 router.get('/',controllersthomas.declarefirstmessage)
+router.post('/test/',controllersthomas.createObject)
 module.exports = router
