+++ back/serveur.js
@@ -1,27 +1,47 @@
-// import dotenv from 'dotenv'
-require('dotenv').config({path:'./.env'})
-const express = require('express')
-let app = express()
-let cors = require('cors')
-// const testroute = require('./routes/testroute.js')
-const connexion = require('./sql/connexion.js')
-app.use(cors())
+const http = require('http');
+const app = require('./app');
 
-app.use("/test", express.static("./"));
-app.use(express.json())
-// app.use("/test",testroute)
-app.get('/connexion', (req, res) => {
-    connexion.query('SELECT * FROM test', (error, results) => {
-      if (error) {
-        console.error('Erreur lors de l\'exécution de la requête :', error);
-        res.status(500).send('Erreur du serveur');
-      } else {
-        res.json(results);
-      }
-    });
-  });
+const normalizePort = val => {
+  const port = parseInt(val, 10);
 
-app.listen(process.env.PORT, () => {
-    console.log(`Server app listening on port ${process.env.PORT}`);
+  if (isNaN(port)) {
+    return val;
+  }
+  if (port >= 0) {
+    return port;
+  }
+  return false;
+};
+const port = normalizePort(process.env.PORT || '3000');
+app.set('port', port);
+
+const errorHandler = error => {
+  if (error.syscall !== 'listen') {
+    throw error;
+  }
+  const address = server.address();
+  const bind = typeof address === 'string' ? 'pipe ' + address : 'port: ' + port;
+  switch (error.code) {
+    case 'EACCES':
+      console.error(bind + ' requires elevated privileges.');
+      process.exit(1);
+      break;
+    case 'EADDRINUSE':
+      console.error(bind + ' is already in use.');
+      process.exit(1);
+      break;
+    default:
+      throw error;
+  }
+};
+
+const server = http.createServer(app);
+
+server.on('error', errorHandler);
+server.on('listening', () => {
+  const address = server.address();
+  const bind = typeof address === 'string' ? 'pipe ' + address : 'port ' + port;
+  console.log('Listening on ' + bind);
 });
-module.exports = app
+
+server.listen(port);
